Move auto_fire_enabled to its own config; fix 3 production bugs; accounts UI to htmx
Two batches of work landed together in this working tree:

1. /admin/accounts' Tạm dừng/Kích hoạt lại/Xoá buttons now use htmx
   (hx-post targeting #accounts-content), matching /admin/groups and
   /admin/schedule — updates the account table in place instead of a
   full page reload, with the existing plain-form fallback kept for when
   htmx isn't available.

2. Real bugs found and fixed 2026-09-07 (not this session's work, but
   verified and committed together per request):
   - auto_fire_enabled moved out of DataSyncConfig into its own
     human_bot/scheduling_config.py: it gates EVERY due scheduled task
     (side-B auto-scheduled or composed by hand at /admin/post), not
     just ones the data-sync poller creates, so burying it under "Đồng
     bộ dữ liệu bên B" made it easy to miss — a manually-scheduled post
     sat past its due time because of exactly this. runtime_config.py
     falls back to a value saved under the old "data_sync" key so an
     existing choice isn't silently reset; .env reads the new
     SCHEDULING_AUTO_FIRE_ENABLED name with the old
     DATA_SYNC_AUTO_FIRE_ENABLED as fallback. /admin/post and
     /admin/schedule now show the gate's live on/off state directly.
   - service.py's fire loop no longer depends on DataSyncConfig.enabled
     — that flag only gates the side-B poller, but was accidentally
     also blocking manually-scheduled posts from ever firing whenever
     side-B sync was turned off.
   - admin.py's /admin/config save handler cast every numeric field to
     float regardless of the dataclass's actual type, silently turning
     int fields (e.g. HumanMouseConfig.min_steps/max_steps) into floats
     — which later crashed human_mouse_move()'s range() with "'float'
     object cannot be interpreted as an integer" on specific click
     distances. Now casts to each field's real declared type.
   - browser_pool.py's AccountSession could get stuck forever pointing
     at a dead page/browser (most commonly: a human closing the visible
     Chrome window under HEADLESS=false) — ensure_started() only checked
     "is it None", not "is it actually still alive", so every task kept
     failing until the whole service was restarted. Added a real
     liveness check with automatic teardown-and-relaunch.

README.md gained a "Điểm yếu đã ghi nhận" section documenting these
fixes plus other known gaps (no post-success verification, no selector
fallback, 4 stub actions, no anti-fingerprinting).

+++ human_bot/humanize.py
@@ -439,7 +439,13 @@ async def human_mouse_move(
         end_x = target_x + (target_x - start_x) * cfg.overshoot_ratio
         end_y = target_y + (target_y - start_y) * cfg.overshoot_ratio
 
-    steps = max(cfg.min_steps, min(cfg.max_steps, round(distance / 25)))
+    # int(...) matters here, not just cosmetic: cfg.min_steps/max_steps are
+    # floats (admin-editable via /admin/config), so whenever the clamp
+    # picks one of those bounds instead of round(distance / 25), max()/min()
+    # hand back that float as-is — range() below then raises "'float'
+    # object cannot be interpreted as an integer" (hit in production
+    # 2026-09-07, only on distances short/long enough to actually clamp).
+    steps = int(max(cfg.min_steps, min(cfg.max_steps, round(distance / 25))))
     for i in range(1, steps + 1):
         t = i / steps
         x, y = _quadratic_bezier((start_x, start_y), (control_x, control_y), (end_x, end_y), t)
